remove some commented code and renamed subfolders in pages folder
