--- project/src/mock/comments.ts
@@ -1,117 +1,0 @@
-import { Comments } from '../types/comment';
-
-const comments: Comments = [
-  {
-    id: 1,
-    user: {
-      id: 13,
-      name: 'Zak'
-    },
-    rating: 5.9,
-    comment: 'This movie is just plain bad. There must be some big payola going round this awards season. Badly written, average acting at best, all the characters are unrelatable and inlikeable. 2 hours of my life wasted.',
-    date: '2022-06-09T12:25:36.946Z'
-  },
-  {
-    id: 2,
-    user: {
-      id: 14,
-      name: 'Corey'
-    },
-    rating: 1.5,
-    comment: 'A movie that will take you to another world full of emotions.',
-    date: '2022-06-05T12:25:36.946Z'
-  },
-  {
-    id: 3,
-    user: {
-      id: 14,
-      name: 'Corey'
-    },
-    rating: 8.6,
-    comment: 'This movie is just plain bad. There must be some big payola going round this awards season. Badly written, average acting at best, all the characters are unrelatable and inlikeable. 2 hours of my life wasted.',
-    date: '2022-06-07T12:25:36.946Z'
-  },
-  {
-    id: 4,
-    user: {
-      id: 13,
-      name: 'Zak'
-    },
-    rating: 5.9,
-    comment: 'This movie is just plain bad. There must be some big payola going round this awards season. Badly written, average acting at best, all the characters are unrelatable and inlikeable. 2 hours of my life wasted.',
-    date: '2022-06-09T12:25:36.946Z'
-  },
-  {
-    id: 5,
-    user: {
-      id: 14,
-      name: 'Corey'
-    },
-    rating: 1.5,
-    comment: 'A movie that will take you to another world full of emotions.',
-    date: '2022-06-05T12:25:36.946Z'
-  },
-  {
-    id: 6,
-    user: {
-      id: 14,
-      name: 'Corey'
-    },
-    rating: 8.6,
-    comment: 'This movie is just plain bad. There must be some big payola going round this awards season. Badly written, average acting at best, all the characters are unrelatable and inlikeable. 2 hours of my life wasted.',
-    date: '2022-06-07T12:25:36.946Z'
-  },
-  {
-    id: 7,
-    user: {
-      id: 14,
-      name: 'Corey'
-    },
-    rating: 8.6,
-    comment: 'This movie is just plain bad. There must be some big payola going round this awards season. Badly written, average acting at best, all the characters are unrelatable and inlikeable. 2 hours of my life wasted.',
-    date: '2022-06-07T12:25:36.946Z'
-  },
-  {
-    id: 8,
-    user: {
-      id: 13,
-      name: 'Zak'
-    },
-    rating: 5.9,
-    comment: 'This movie is just plain bad. There must be some big payola going round this awards season. Badly written, average acting at best, all the characters are unrelatable and inlikeable. 2 hours of my life wasted.',
-    date: '2022-06-09T12:25:36.946Z'
-  },
-  {
-    id: 9,
-    user: {
-      id: 14,
-      name: 'Corey'
-    },
-    rating: 1.5,
-    comment: 'A movie that will take you to another world full of emotions.',
-    date: '2022-06-05T12:25:36.946Z'
-  },
-  {
-    id: 10,
-    user: {
-      id: 14,
-      name: 'Corey'
-    },
-    rating: 8.6,
-    comment: 'This movie is just plain bad. There must be some big payola going round this awards season. Badly written, average acting at best, all the characters are unrelatable and inlikeable. 2 hours of my life wasted.',
-    date: '2022-06-07T12:25:36.946Z'
-  },
-  {
-    id: 11,
-    user: {
-      id: 14,
-      name: 'Corey'
-    },
-    rating: 8.6,
-    comment: 'This movie is just plain bad. There must be some big payola going round this awards season. Badly written, average acting at best, all the characters are unrelatable and inlikeable. 2 hours of my life wasted.',
-    date: '2022-06-07T12:25:36.946Z'
-  },
-
-];
-
-export default comments;
