+++ project/src/pages/not-found-page/not-found-page.tsx
@@ -1,6 +1,5 @@
 import { Link } from 'react-router-dom';
 import Logo from '../../components/logo/logo';
-import SomeComp from '../../components/some-comp/some-comp';
 import { LogoState } from '../../util/const';
 
 function NotFoundPage(): JSX.Element {
@@ -8,7 +7,6 @@ function NotFoundPage(): JSX.Element {
     <div className="not-found-page">
       <header className="page-header">
         <Logo renderPlace={LogoState.Header}/>
-        <SomeComp addElement={false} />
       </header>
 
       <div className="not-found-page">
